requirements.txt
Django>=4.2.8,<4.3.0
djangorestframework>=3.14.0,<3.15
git+https://github.com/rholdxz/django-jazzmin.git@b55c65588cb16e6e9d75e87dc70d9156bde00109
psycopg2>=2.9.9,<3.0
django-phonenumber-field[phonenumbers]>=7.2.0,<7.3.0